Initial RSA encode and Decode
Right now I take an EditText and the 'message' in there is run through
RSA with new generating keys every time. Will need to write the KeyPair
to memory and then check if they are there or generate once and use a
shared preference to say I have a keypair stored (probably faster).